all up to date 30.12.2020

+++ Controllers/DeleteSongController.cs
@@ -4,8 +4,12 @@
 * This program is not allowed to be copied or reused without explicit permission.
 * erik10cavazos@gmail.com and everardo.cavazoshrnnd@uanl.edu.mx
 */
+
+using System;
 using Microsoft.AspNetCore.Mvc;
 using isolaatti_API.Models;
+using Microsoft.AspNetCore.Html;
+using Microsoft.AspNetCore.Mvc.RazorPages;
 
 namespace isolaatti_API.Controllers
 {
@@ -20,16 +24,27 @@ namespace isolaatti_API.Controllers
         }
         // normal delete. includes: deleting song record from db, and deleting the files from server
         [HttpPost]
-        public string Index([FromForm] int songId)
+        public IActionResult Index([FromForm] int songId, [FromForm] int userId, [FromForm] string password)
         {
-            Song songToDelete = db.Songs.Find(songId);
+            try
+            {
+                User user = db.Users.Find(userId);
+                if (user.Password.Equals(password))
+                {
+                    Song songToDelete = db.Songs.Find(songId);
+                
+                    // deletes database record of song
+                    db.Songs.Remove(songToDelete);
+                    db.SaveChanges();
+                    return Ok(songToDelete.Uid);
+                }
+            }
+            catch (InvalidOperationException)
+            {
+                return StatusCode(404);
+            }
 
-            string uid = songToDelete.BassUrl.Split("/")[4];
-            // deletes database record of song
-            db.Songs.Remove(songToDelete);
-            db.SaveChanges();
-
-            return uid;
+            return StatusCode(403);
         }
     }
 }
